mirror-sdk-py/pyproject.toml
[project]
name = "ocx-gen"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["github3.py>=4", "httpx>=0.28"]

[project.optional-dependencies]
dev = ["datamodel-code-generator>=0.26", "pytest>=8", "ruff>=0.9"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ocx_gen"]
